Move right hand most popular out of experiments folder
