+++ common/app/assets/javascripts/modules/onward/right-most-popular-image-item.js
@@ -3,11 +3,9 @@
  Description: Item for right most popular panel
  */
 define([
-    'common/modules/component',
-    'bonzo'
+    'common/modules/component'
 ], function (
-    Component,
-    bonzo
+    Component
     ) {
 
     function RightMostPopularItem(data, index) {
@@ -21,19 +19,27 @@ define([
     RightMostPopularItem.prototype.componentClass = 'right-most-popular-item';
     RightMostPopularItem.prototype.classes = {
         link: 'url',
-        count: 'count',
+        image: 'img',
         headline: 'headline' };
     RightMostPopularItem.prototype.useBem = true;
+    RightMostPopularItem.prototype.showComments = false;
 
     RightMostPopularItem.prototype.template = '<li class="right-most-popular-item"><a class="right-most-popular-item__url media u-cf" href="">'
-        + '<div class="right-most-popular-item__count-container media__img"><span class="right-most-popular-item__count"></span></div>'
-        + '<h3 class="right-most-popular-item__headline media__body"></h3></a></li>';
+        + '<div class="right-most-popular-item__img media__img js-image-upgrade"><img class="responsive-img" src="" alt=""/></div>'
+        + '<h3 class="right-most-popular-item__headline media__body">'
+        + '</h3></a></li>';
 
     RightMostPopularItem.prototype.prerender = function() {
-        this.getElem(this.classes.count).innerHTML =  this.index;
+        if (this.data.itemPicture) {
+            var container = this.getElem(this.classes.image);
+            container.setAttribute("data-src", this.data.itemPicture);
+        }
+        if(this.data.discussionId && this.showComments) {
+            this.elem.setAttribute("data-discussion-id", this.data.discussionId);
+        }
+        this.elem.setAttribute('data-link-name', 'trail | ' + this.index);
         this.getElem(this.classes.link).href = this.data.url;
         this.getElem(this.classes.headline).innerHTML = this.data.headline;
-        this.elem.setAttribute('data-link-name', 'trail | ' + this.index);
     };
 
     return RightMostPopularItem;
