+++ common/app/assets/stylesheets/global.scss
@@ -50,6 +50,7 @@
 @import "module/facia/_mixins";
 @import "module/containers/_popular";
 @import "module/onward/_right-ear";
+@import "module/onward/_right-hand-most-popular";
 @import "module/_social";
     @include socialDelayed;
 @import "module/_clamp";
@@ -72,7 +73,6 @@
 
 @import "module/experiments/_ab-inline-link-card";
 @import "module/experiments/_highlight-panel";
-@import "module/experiments/_right-hand-most-popular";
 @import "module/experiments/_underline-links";
 
 /* ==========================================================================
