+++ common/app/assets/stylesheets/module/onward/_right-hand-most-popular.scss
